[DEV]Added rectangle & vec2 seperate & nArr bound
Added the Vector 2 seperate function to get an array of x and y coords.
Added the numberArray isBounded function to know if the values are
not extremums of the array.
Added the rectangle class along with the contains function to know if
a rect is inside another & basic rectangle calculations